Образование контроллера Issue, создание отдельных Views
