+++ FinDesk2/Views/Shared/Partial/_MainMenu.cshtml
@@ -17,7 +17,7 @@
         <li><a asp-controller="Home" asp-action="Error404">404</a></li>
         <li><a asp-controller="Home" asp-action="ContactUs">Контакты</a></li>
         <li><a asp-controller="Employees" asp-action="Index">Сотрудники</a></li>
-        <li><a asp-controller="Home" asp-action="Issues">Issues</a></li>
+        <li><a asp-controller="Issue" asp-action="Index">Issues</a></li>
 
     </ul>
 </div>
